+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [2.5.3]
+
+### Added
+
+- Added support for custom title formatting: Doublespace and Monospace
+  - Chose to set the focussed window title to monospace format
+
+### Fixed
+
+- The fix applied in 2.5.2 – apply it to the global title as well
+
 ## [2.5.2]
 
 ### Fixed
@@ -852,7 +863,8 @@
 - Window color management
 - Title numbering options
 
-[Unreleased]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.5.2...HEAD
+[Unreleased]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.5.3...HEAD
+[2.5.3]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.5.2...2.5.3
 [2.5.2]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.5.1...2.5.2
 [2.5.1]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.5.0...2.5.1
 [2.5.0]: https://github.com/alexBlakeGoudemond/jetbrains-window-accent/compare/2.4.1...2.5.0